+++ src/main/java/com/autohome/Main.java
@@ -20,16 +20,4 @@ public class Main {
         LOG.info("JVM shutting down");
     }
 
-/*
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http.authorizeRequests(a -> a.antMatchers("/", "/ping", "/error").permitAll()
-                .anyRequest().authenticated())
-                .exceptionHandling(e -> e.authenticationEntryPoint(new HttpStatusEntryPoint(HttpStatus.UNAUTHORIZED)))
-                .logout(l -> l.logoutSuccessUrl("/").permitAll())
-                .csrf(c -> c.csrfTokenRepository(CookieCsrfTokenRepository.withHttpOnlyFalse()))
-                .httpBasic();
-    }
-*/
-
 }
