+++ src/main/resources/application.properties
@@ -5,17 +5,3 @@ logging.file=autohome.logs
 #Sql
 spring.datasource.url=jdbc:mysql://192.168.1.250:3306/autohome
 spring.datasource.driverclassname = com.mysql.cj.jdbc.Driver
-
-spring.security.oauth2.client.registration.google.clientId=750505933588-ns93dn1hlkfj54bu6e7eg4p9m8hckb40.apps.googleusercontent.com
-spring.security.oauth2.client.registration.google.clientSecret=pRZrbv64X4WtZHhQo30sZHWd
-#security.oauth2.client.project_id=zinc-prism-275020
-#security.oauth2.client.userAuthorizationUri=https://accounts.google.com/o/oauth2/auth
-#security.oauth2.client.accessTokenUri=https://oauth2.googleapis.com/token
-#security.oauth2.client.auth_provider_x509_cert_url=https://www.googleapis.com/oauth2/v1/certs
-#security.oauth2.client.tokenName=oauth_token
-#security.oauth2.client.authenticationScheme=query
-#security.oauth2.client.clientAuthenticationScheme=form
-#security.oauth2.client.scope=profile email
-#
-#security.oauth2.resource.userInfoUri=https://www.googleapis.com/userinfo/v2/me
-#security.oauth2.resource.preferTokenInfo=false
